fix(styles): media queries for widget editor and controls
Fixed widths seem to have no effect on high resolutions, but in low resolutions it makes the
"Selected Properties" section of the Widgets Editor to have invisible or out of place content,
including the scrollbar, color picker, and probably others.

I've added media queries to fix the particular case of low resolution, but we should consider
removing the fixed widths.